# apiDados/swagger.json
{
    "swagger": "2.0",
    "info": {
        "version": "1.0.0",
        "title": "Api Hypatiamat",
        "description": "Aqui encontra-se a documentação de toda a api de dados do Hypatiamat, incluindo todas as rotas e toda a informação à cerca do que é precisar passar como parâmetros, bem como os resultados que se podem esperar ao executar aquela rota."
    },
    "schemes": [
        "http"
    ],
    "host": "localhost:3050",
    "basePath": "/",
...
    "paths": {
...
          "/aplicacoes/alunos": {
            "get": {
                "tags":[
                    "Alunos"
                ],
                "summary": "Todos os alunos registados.",
                "description": "Retorna uma lista com a informação à cerca de todos os alunos.",
                "responses": {
                    "200": {
                        "description": "Lista de Alunos",
                        "schema": {
                          "type":"array",
                          "items":{
                            "$ref": "#/definitions/Aluno"
                          }
                        }
                    },
                    "401":{
                      "description": "Falta-lhe o token de acesso ou o token não é válido."
                    }
                },
                "security": [
                  {
                    "api_key": []
                  }
                ]
            },
            "post":{
                "tags":[
                    "Alunos"
                ],
                "summary": "Inserir um novo aluno.",
                "description": "Adiciona um novo aluno à base de dados.",
                "parameters": [
                    {
                      "name": "user",
                      "in": "formData",
                      "description": "Código/Username do aluno",
                      "required": true,
                      "type": "string"
                    },
                    {
                      "name": "numero",
                      "in": "formData",
                      "description": "Número do aluno",
                      "required": true,
                      "type": "integer"
                    },
                    {
                      "name": "nome",
                      "in": "formData",
                      "description": "Nome do aluno",
                      "required": true,
                      "type": "string"
                    },
                    {
                      "name": "datanascimento",
                      "in": "formData",
                      "description": "Data de nascimento do aluno",
                      "required": true,
                      "type": "integer"
                    },
                    {
                      "name": "escola",
                      "in": "formData",
                      "description": "Identificador da escola do aluno",
                      "required": true,
                      "type": "integer"
                    },
                    {
                      "name": "turma",
                      "in": "formData",
                      "description": "Turma do aluno",
                      "required": true,
                      "type": "string"
                    },
                    {
                      "name": "email",
                      "in": "formData",
                      "description": "Email do aluno",
                      "required": true,
                      "type": "string"
                    },{
                      "name": "password",
                      "in": "formData",
                      "description": "Password do aluno",
                      "required": true,
                      "type": "string",
                      "format": "password"
                    },{
                      "name": "codprofessor",
                      "in": "formData",
                      "description": "Código de acesso fornecido pelo professor",
                      "required": true,
                      "type": "string"
                    },{
                      
                      "name": "pais",
                      "in": "formData",
                      "description": "País do aluno",
                      "required": true,
                      "type": "string"
                    },{
                      "name": "confirmacao",
                      "in": "formData",
                      "description": "1 caso esteja confirmado, 0 caso contrário",
                      "required": true,
                      "type": "integer"
                    }
                  ],
                "responses": {
                    "200": {
                        "description": "Retorna o id do aluno criado",
                        "schema": {
                            "type": "string",
                            "items": {
                                "properties": {
                                    "id": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/aplicacoes/alunos/{idAluno}": {
            "get": {
              "tags": [
                "Alunos"
              ],
              "summary": "Dado um identifcador de aluno, devolve toda a informação relativa a esse aluno.",
              "description": "Devolve toda a informação relativa a um professor.",
              "operationId": "getAlunoById",
              "parameters": [
                {
                  "name": "idAluno",
                  "in": "path",
                  "description": "Identificador do aluno a retornar.",
                  "required": true,
                  "type": "integer"
                }
              ],
              "responses": {
                "200": {
                    "description": "Retorna a informação relativa a um aluno.",
                    "schema": {
                        "$ref": "#/definitions/Aluno"   
                        }
                },
                "401":{
                  "description": "Falta-lhe o token de acesso ou o token não é válido."
                },
                "403":{
                  "description": "Não possuí nível de acesso suficiente."
                }
              },
              "security": [
                {
                  "api_key": []
                }
              ]
            },
            "put":{
              "tags": [
                "Alunos"
              ],
              "summary": "Dado um identificador de um aluno, edite a informação relacionada com ele.",
              "description": "Editar um aluno.",
              "operationId": "putAlunoById",
              "parameters": [
                {
                  "name": "idAluno",
                  "in": "path",
                  "description": "Identificador de um aluno.",
                  "required": true,
                  "type": "integer"
                },
                {
                  "name": "numero",
                  "in": "formData",
                  "description": "Número do aluno",
                  "required": true,
                  "type": "integer"
                },
                {
                  "name": "nome",
                  "in": "formData",
                  "description": "Nome do aluno",
                  "required": true,
                  "type": "string"
                },
                {
                  "name": "datanascimento",
                  "in": "formData",
                  "description": "Data de nascimento do aluno",
                  "required": true,
                  "type": "integer"
                },
                {
                  "name": "email",
                  "in": "formData",
                  "description": "Email do aluno",
                  "required": true,
                  "type": "string"
                },
                {
                  "name": "confirmacao",
                  "in": "formData",
                  "description": "1 caso esteja confirmado, 0 caso contrário",
                  "required": true,
                  "type": "integer"
                }
              ],
              "responses": {
                "200": {
                  "description": "O aluno foi editado."
                },
                "401":{
                  "description": "Falta-lhe o token de acesso ou o token não é válido."
                },
                "403":{
                  "description": "Não possuí nível de acesso suficiente."
                }
              },
                "security": [
                  {
                    "api_key": []
                  }
                ]
            },
            "delete": {
              "tags": [
                "Alunos"
              ],
              "summary": "Dado um identificador de aluno, elimine o aluno da base de dados.",
              "description": "Apagar um aluno.",
              "operationId": "deleteAlunoById",
              "parameters": [
                {
                  "name": "idAluno",
                  "in": "path",
                  "description": "Identificador do aluno a apagar.",
                  "required": true,
                  "type": "string"
                }
              ],
              "responses": {
                "200": {
                  "description": "O aluno foi eliminado."
                },
                "401":{
                  "description": "Falta-lhe o token de acesso ou o token não é válido."
                },
                "403":{
                  "description": "Não possuí nível de acesso suficiente."
                }
              },
              "security": [
                {
                  "api_key": []
                }
              ]
              
            }
        },
        "/aplicacoes/alunos/{idAluno}/password":{
          "put":{
            "tags": [
              "Alunos"
            ],
            "summary": "Dado um identificador de um aluno, edite a password dele.",
            "description": "Editar a password de um aluno.",
            "operationId": "putPasswordAlunoById",
            "parameters": [
              {
                "name": "idAluno",
                "in": "path",
                "description": "Identificador de um aluno.",
                "required": true,
                "type": "integer"
              },
              {
                "name": "password",
                "in": "formData",
                "description": "Nova password do aluno",
                "required": true,
                "type": "string",
                "format": "password"
              }
            ],
            "responses": {
              "200": {
                "description": "A palavra passe do aluno foi alterada."
              },
              "401":{
                "description": "Falta-lhe o token de acesso ou o token não é válido."
              },
              "403":{
                "description": "Não possuí nível de acesso suficiente."
              }
            },
              "security": [
                {
                  "api_key": []
                }
              ]
          }
        },
        "/aplicacoes/alunos/{idAluno}/escola":{
          "put":{
            "tags": [
              "Alunos"
            ],
            "summary": "Dado um identificador de um aluno, edite a escola dele.",
            "description": "Editar a escola de um aluno.",
            "operationId": "putEscolaAlunoById",
            "parameters": [
              {
                "name": "idAluno",
                "in": "path",
                "description": "Identificador de um aluno.",
                "required": true,
                "type": "integer"
              },
              {
                "name": "escola",
                "in": "formData",
                "description": "Nova escola do aluno (identificador da escola)",
                "required": true,
                "type": "integer"
              }
            ],
            "responses": {
              "200": {
                "description": "A escola do aluno foi alterada."
              },
              "401":{
                "description": "Falta-lhe o token de acesso ou o token não é válido."
              },
              "403":{
                "description": "Não possuí nível de acesso suficiente."
              }
            },
              "security": [
                {
                  "api_key": []
                }
              ]
          }
        },
        "/aplicacoes/alunos/{idAluno}/turma":{
          "put":{
            "tags": [
              "Alunos"
            ],
            "summary": "Dado um identificador de um aluno, edite a password dele.",
            "description": "Editar a password de um aluno.",
            "operationId": "putPasswordAlunoById",
            "parameters": [
              {
                "name": "idAluno",
                "in": "path",
                "description": "Identificador de um aluno.",
                "required": true,
                "type": "integer"
              },
              {
                "name": "turma",
                "in": "formData",
                "description": "Nova turma do aluno.",
                "required": true,
                "type": "string"
              }
            ],
            "responses": {
              "200": {
                "description": "A turma do aluno foi alterada."
              },
              "401":{
                "description": "Falta-lhe o token de acesso ou o token não é válido."
              },
              "403":{
                "description": "Não possuí nível de acesso suficiente."
              }
            },
              "security": [
                {
                  "api_key": []
                }
              ]
          }
        },
...
      "/aplicacoes/escolas/{idEscola}/alunos":{
        "get": {
          "tags": [
            "Escolas"
          ],
          "summary": "Dado um identifcador da escola, devolve todos os alunos da escola.",
          "description": "Devolve todos os alunos de uma escola.",
          "operationId": "getAlunosByEscola",
          "parameters": [
            {
              "name": "idEscola",
              "in": "path",
              "description": "Identificador da escola.",
              "required": true,
              "type": "integer"
            }
          ],
          "responses": {
            "200": {
                "description": "Retorna todos os alunos da escola.",
                "schema": {
                    "type":"array",
                    "items": {
                      "$ref": "#/definitions/Aluno"
                    }   
                    }
            },
            "401":{
              "description": "Falta-lhe o token de acesso ou o token não é válido."
            },
            "403":{
              "description": "Não possuí nível de acesso suficiente."
            }
          },
          "security": [
            {
              "api_key": []
            }
          ]
        }
      },
...
      "/aplicacoes/turmas/{turma}/alunos":{
        "get": {
          "tags": [
            "Turmas"
          ],
          "summary": "Dado um identifcador da turma, devolve todos os seus alunos .",
          "description": "Devolve todos os alunos de uma turma.",
          "operationId": "getAlunosByTurma",
          "parameters": [
            {
              "name": "turma",
              "in": "path",
              "description": "Nome da turma.",
              "required": true,
              "type": "string"
            }
          ],
          "responses": {
            "200": {
                "description": "Lista de alunos da turma.",
                "schema": {
                    "type":"array",
                    "items": {
                      "$ref": "#/definitions/Aluno"
                    }   
                }
            },
            "401":{
              "description": "Falta-lhe o token de acesso ou o token não é válido."
            },
            "403":{
              "description": "Não possuí nível de acesso suficiente."
            }
          },
          "security": [
            {
              "api_key": []
            }
          ]
        }
      }
    },
    "securityDefinitions":{
      "api_key":{
        "type": "apiKey",
        "in": "query",
        "name": "token"
      }
    },
    "security": [{
       "api_key": []
    }],
    "definitions": {
        "Aluno": {
            "type": "object",
            "properties": {
              "id": {
                "type": "integer"
              },
              "user": {
                "type": "string"
              },
              "numero": {
                "type": "integer"
              },
              "nome": {
                "type": "string"
              },
              "datanascimento": {
                "type": "string"
              },
              "escola": {
                "type": "string"
              },
              "turma":{
                "type": "string"
              },
              "email":{
                "type":"string"
              },
              "confirmacao":{
                "type":"integer"
              }
            }
        },
...
    }   
}